requirements.txt
streamlit
opencv-python-headless
ultralytics
Pillow
numpy
torch
